requirements.txt
aniso8601==6.0.0
certifi==2018.11.29
chardet==3.0.4
python-decouple>=3.1,<3.99
python-dotenv==0.10.1
future==0.17.1
future-fstrings
idna==2.8
-e git+https://github.com/IMSGlobal/caliper-python.git@a8549ae7227cef896f222626ffd15c4ef4bf9edf#egg=imsglobal-caliper
psycopg2==2.7.7
requests==2.21.0
rfc3986==1.3.2
tokenize-rt==2.1.0
urllib3==1.24.2
python-dateutil==2.8.1
